sanitycheck: tool to update testsuite
New tool is created to use with Testrail.
Purpose of that script is to update a test suite in the Testrail,
by uploading test cases metedata using the source .csv file.
(source file is generating by --export-tests sanitycheck option)
1. That tool detects if new sections was added in the source code,
and make them aligned with the test suite in the Testrail system.
2. Script detects if new test cases were added in the source code,
and uploads new test cases to the Testrail.
After run of that tool test suite of the source code
and test suite of the Testrail will be aligned with each other.
Then user can run script report.py to upload results to the Testrail.

Signed-off-by: Maksim Masalski <maksim.masalski@intel.com>